requirements.txt
Django==2.1.5
django-ckeditor==5.3.0
django-crispy-forms==1.8.0
djangorestframework==3.11.0
Pillow==4.3.0
